refactor: extract SystemAudioFileWriter from SystemAudioRecorder (#548, #435)
Second slice of #435, decomposing the CoreAudio god object. Moves the
recording file writer/finalizer — the AVAudioFile-wrapping
SystemAudioFileWriter (internal, @unchecked Sendable) — out of
SystemAudioRecorder.swift into its own file, byte-identical (class body
and thread-safety comment unchanged; the new file adds only its
file-local imports). No CoreAudio, recorder, or private-extension logic
changed; SystemAudioRecorder and SystemAudioTapSession still reference
it cross-file (same module, internal).

The existing format-invalidation test moves out of
RecordingAudioSourceTests into a new focused, permission-free
SystemAudioFileWriterTests (no duplicate coverage left behind), joined
by lifecycle + edge tests: init creates a readable AVAudioFile; clean
close() finalizes without throwing; markFormatInvalidated + close()
throws recordingFailure; a second close() after either a clean or an
invalidated close() does not re-throw (handle released, flags reset
under the lock); init under a nonexistent parent directory throws; and a
unicode/emoji filename still initializes and finalizes.

Per the codex pre-build review, the formatInvalidated-before-writeError
precedence stays a source-reviewed invariant, not a unit test: writeError
is private and only set by a real AVAudioFile write failure, which cannot
be induced permission-free without a test hook or visibility widening
(both out of scope). 662 unit tests pass; docs-sync validation passes.

Pre-build spec reviewed by codex (5 findings folded in before build).
