+++ Tests/BugNarratorTests/RecordingAudioSourceTests.swift
@@ -32,27 +32,6 @@ final class RecordingAudioSourceTests: XCTestCase {
         XCTAssertFalse(SystemAudioAggregateDeviceIdentity.isOwnedAggregateDeviceUID(""))
     }
 
-    func testSystemAudioFileWriterFailsCloseWhenFormatInvalidated() throws {
-        let fileURL = FileManager.default.temporaryDirectory
-            .appendingPathComponent("system-audio-format-invalidated-\(UUID().uuidString)")
-            .appendingPathExtension("wav")
-        defer { try? FileManager.default.removeItem(at: fileURL) }
-
-        let format = try XCTUnwrap(AVAudioFormat(standardFormatWithSampleRate: 44_100, channels: 2))
-        let writer = try SystemAudioFileWriter(fileURL: fileURL, format: format)
-
-        writer.markFormatInvalidated()
-
-        XCTAssertThrowsError(try writer.close()) { error in
-            guard case let AppError.recordingFailure(message) = error else {
-                XCTFail("Expected recordingFailure, got \(error).")
-                return
-            }
-
-            XCTAssertTrue(message.contains("System audio format changed while recording"))
-        }
-    }
-
     func testMicrophoneLevelCalculatorReturnsZeroForSilentBuffer() throws {
         let buffer = try makePCMBuffer(samples: [0, 0, 0, 0])
